Updated location checking
# TODO taken care of:
function now checks if the song exists in the target directory. If the
song exists in the target directory, then play it.

+++ SongManipulation/play_song.py
@@ -2,6 +2,7 @@
 from gmusicapi import Mobileclient
 from gmusicapi import Musicmanager
 import re
+import os.path
 
 # import vlc
 
@@ -31,11 +32,19 @@ def play_song(song="Raindrop"):
                     print 'Song found!'
                     song_id = song['id']
                     filename, audio = mm.download_song(song_id)
-                    # TODO
-                    # check if song already exists in song_location, no need to download again
-                    # just play
-                    with open(song_location + filename, 'wb') as f:
-                        f.write(audio)
+                    # check if song is already downloaded
+                    try:
+                        print song_location + filename
+                        if os.path.isfile(song_location + filename):
+                            print 'Song is already downloaded...'
+                            print 'Playing song.'
+                            break
+                        else:
+                            with open(song_location + filename, 'wb') as f:
+                                f.write(audio)
+                    except (OSError, IOError):
+                        print 'An error has occurred.'
+                        break
 
                     # TODO
                     # vlc.MediaPlayer.play(filename)
